+++ src/lib/admin/LauncherInformation.svelte
