+++ src/lib/common/ManualMenu.svelte
