+++ src/routes/[lang=lang]/(manual)/manual/[maindir]/[subdir]/+page.svelte
@@ -1,5 +1,5 @@
 <script lang="ts">
-    import Menu from '$lib/common/Menu.svelte';
+    import Menu from '$lib/common/ManualMenu.svelte';
     import LL from '$i18n/i18n-svelte';
 
     export let data;
