+++ static/ts/main.ts
@@ -1,4 +1,5 @@
 import { browser } from '$app/environment';
+import { writable } from 'svelte/store'
 
 /*=========================================================
 　　　　　Slide Functions
@@ -136,6 +137,11 @@ if (browser) {
         });
     });
 }
+
+/*=========================================================
+　　　　　Admin Console
+=======================================================*/
+export const tabParam = writable('')
 
 /*=========================================================
 　　　　　Misc
